Rectification d'une transformation en int redondante

+++ src/main/java/fr/univartois/butinfo/lensymphony/notes/FermataNote.java
@@ -22,6 +22,6 @@ public class FermataNote extends NoteDecorator {
      */
     @Override
     public int getDuration(int tempo) {
-        return (int) (super.getDuration(tempo) * 4); // Fermata note during is 4 times the original note duration
+        return super.getDuration(tempo) * 4; // Fermata note during is 4 times the original note duration
     }
 }
